Add flash attention support (hardcoded) and make compatible nbafl trainer by fixing the dtype of tensor of ones to float16

+++ federatedscope/core/trainers/trainer_nbafl.py
@@ -138,7 +138,7 @@ def inject_noise_in_broadcast(cfg, sample_client_num, model):
     # Clip weight
     for p in model.parameters():
         p.data = p.data / torch.max(
-            torch.ones(size=p.shape, device=p.data.device),
+            torch.ones(size=p.shape, device=p.data.device, dtype=torch.float16),
             torch.abs(p.data) / cfg.nbafl.w_clip)
     if len(sample_client_num) > 0:
         # Inject noise
